Refactor core traits Message, ClassType and ProtocolType
