+++ crates/test-ui/ui/mutable_id_not_clone_not_retain.stderr
@@ -39,7 +39,7 @@ error[E0277]: the trait bound `Mutable: mutability::MutabilityIsRetainable` is n
             MainThreadOnly
   = note: required for `NSMutableObject` to implement `IsRetainable`
 note: required by a bound in `retain`
- --> $WORKSPACE/crates/objc2/src/class_type.rs
+ --> $WORKSPACE/crates/objc2/src/top_level_traits.rs
   |
   |     fn retain(&self) -> Id<Self>
   |        ------ required by a bound in this associated function
